requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
streamlit==1.31.0
sqlmodel==0.0.14
pydantic==2.6.0
python-dotenv==1.0.1
google-generativeai==0.3.2
groq==0.4.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
google-search-results==2.4.2
psycopg2-binary==2.9.9
beautifulsoup4==4.12.3